Extracting EmptyConversationsView
